media-tracks/audiotracks-inband-single-track.html

<!DOCTYPE html>
<head>
  <title>HTMLMediaElement.audioTracks test for inband tracks (single track)</title>
  <script src='resources/testharness.js'></script>
  <script src='resources/testharnessreport.js'></script>
</head>
<script>
// TODO: add '/' before resources
// TODO: use w3c-wpt to get file

const MEDIA_FILE = 'media/test-title-no-lang-no-handler.webm';

promise_test(t => {
  var audio = document.createElement('audio');
  const audioTracks = audio.audioTracks;
  const eventWatcher = new EventWatcher(t, audioTracks, 'addtrack');
  audio.src = MEDIA_FILE;
  return eventWatcher.wait_for('addtrack');
}, 'Test that addtrack event is fired during load.');

promise_test(t => {
  var audio = document.createElement('audio');
  const audioTracks = audio.audioTracks;
  const eventWatcher = new EventWatcher(t, audioTracks, 'change');
  var promise = eventWatcher.wait_for('change');
  audio.src = MEDIA_FILE;
  audio.addEventListener('loadedmetadata', t.step_func(e => {
    audioTracks[0].enabled = false;
  }));
  return promise;
}, 'Test that a change event is fired when the audio track is disabled.');

promise_test(t => {
  var audio = document.createElement('audio');
  const audioTracks = audio.audioTracks;
  const eventWatcher = new EventWatcher(t, audioTracks, 'change');
  var promise = eventWatcher.wait_for([ 'change', 'change' ]);
  audio.src = MEDIA_FILE;
  audio.addEventListener('loadedmetadata', t.step_func(e => {
    audioTracks[0].enabled = false;
    audioTracks[0].enabled = true;
  }));
  return promise;
}, 'Test that a change event is fired when the audio track is re-enabled.');

async_test(t => {
  var audio = document.createElement('audio');
  audio.src = MEDIA_FILE;
  audio.addEventListener('loadedmetadata', t.step_func_done(e => {
    const audioTracks = audio.audioTracks;
    assert_equals(audioTracks.length, 1);
    assert_not_equals(audioTracks[0], null);
    assert_equals(audioTracks[1], undefined);
  }));
}, 'Test that audioTracks has the right length after load.');

async_test(t => {
  var audio = document.createElement('audio');
  audio.src = MEDIA_FILE;
  audio.addEventListener('loadedmetadata', t.step_func_done(e => {
    const audioTracks = audio.audioTracks;
    assert_equals(audioTracks.length, 1);
    assert_equals(audioTracks.getTrackById(0), null);
    assert_not_equals(audioTracks.getTrackById(audioTracks[0].id), null);
  }));
}, 'Test that audioTracks.getTrackById() uses track internal id.');
</script>
